+++ apps/web/types/auth/index.ts
@@ -1,1 +1,4 @@
 //NOTE - Auth 페이지 타입 관리
+
+export type Role = 'BUYER' | 'SELLER';
+export type Provider = 'google' | 'kakao';
